perf(database): remove integer bitwise operations
chore(database): remove TELLY_UNSPECIFIED type
perf(database): add byte count for numbers when saving data to the file
perf(database): remove multiple reading data type
feat(database): change int to long for number data

+++ src/utils/integer.c
@@ -21,7 +21,7 @@ bool is_integer(const char *value) {
   return (_value != value) && (*_value == 0x00);
 }
 
-uint32_t get_digit_count(int32_t number) {
+uint32_t get_digit_count(int64_t number) {
   if (number == 0) {
     return 1;
   } else if (number < 0) {
